Corrected error while exporting timecards and expenses
